RPTS-100 consolidate reusable test data
